requirements.txt
krakenex
pandas
numpy
matplotlib
requests
seaborn
pillow
pykrakenapi